requirements.txt
torch>=2.2
transformers>=4.40
accelerate>=0.28
pillow>=10.0
tqdm>=4.66
